+++ src/main/java/edu/odu/cs/cs330/items/Tool.java
@@ -79,11 +79,11 @@ public class Tool extends Equippable {
         this.setName(tokens[0]);
 
         // Complete this method.
-        this.setDurability(Integer.parseInt(tokens[1]));
-        this.setSpeed(Integer.parseInt(tokens[2]));
+        this.setDurability((int) Integer.parseInt(tokens[1]));
+        this.setSpeed((int) Integer.parseInt(tokens[2]));
         this.setMaterial(tokens[3]);
         this.setModifier(tokens[4]);
-        this.setModifierLevel(Integer.parseInt(tokens[5]));
+        this.setModifierLevel((int) Integer.parseInt(tokens[5]));
     }
 
     /**
